refactor: move module_translator tests to code_translator test,
rename code_translator.rs to mod.rs.
